+++ packData/packData.mjs
@@ -33,7 +33,8 @@ let packs = [
     'cpr-3rd-party-spells',
     'cpr-monster-features-2024',
     'cpr-3rd-party-spells-2024',
-    'cpr-species-features-2024'
+    'cpr-species-features-2024',
+    'cpr-other-features'
 ];
 for (let i of packs) {
     await compilePack('./packData/' + i, './packs/' + i, {log: true});