+++ scripts/macros/2014/piety/karametra/karametrasDevotee.js
@@ -1,0 +1,5 @@
+export let karametrasDevotee = {
+    name: 'Karametra\'s Devotee',
+    version: '1.3.153',
+    rules: 'legacy'
+};