+++ scripts/macros/2014/piety/nylea/nyleasDevotee.js
@@ -1,0 +1,5 @@
+export let nyleasDevotee = {
+    name: 'Nylea\'s Devotee',
+    version: '1.3.153',
+    rules: 'legacy'
+};