+++ scripts/macros/2024/spells/bless.js
@@ -1,7 +1,7 @@
 import {bless as blessLegacy} from '../../2014/spells/bless.js';
 export let bless = {
     name: 'Bless',
-    version: '1.1.17',
+    version: '1.3.153',
     rules: 'modern',
     midi: blessLegacy.midi
 };